pyproject.toml
[tool.poetry]
name = "primeiro-crud-backfront"
version = "0.1.0"
description = ""
authors = ["PauloHBSF <paulohbsfilho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.9.1"}
fastapi = "^0.114.1"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
